+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Changed
 - moved location API url from class Shipment to class PickupPoints
+- changed locations API url to v3
+
+### Added
+- added request error return
 
 ## [2.3.7] - Logo & PDF lib
 ### Updated